Moved GExiv2 specific stuff to a new module
This is a first step towards being able to use an alternative exiv2
library, such as https://github.com/jim-easterbrook/python-exiv2
