+++ src/photini/__init__.py
@@ -1,4 +1,4 @@
 from __future__ import unicode_literals
 
 __version__ = '2021.8.0'
-build = '1791 (bd38c8e)'
+build = '1792 (d17d355)'
